rename layers to modules
