+++ test/3x/torch/quantization/fp8/test_fp8.py
@@ -20,7 +20,7 @@ from neural_compressor.torch.algorithms.habana_fp8.modules import (
 )
 from neural_compressor.torch.quantization import quantize
 from neural_compressor.torch.quantization.config import FP8QConfig, get_default_fp8_qconfig
-from neural_compressor.torch.quantization.layers import BatchMatmul, Matmul
+from neural_compressor.torch.quantization.modules import BatchMatmul, Matmul
 from neural_compressor.torch.utils import logger
 
 torch.set_grad_enabled(False)
