+++ test/requirements.txt
@@ -11,7 +11,6 @@ onnxruntime
 onnxruntime-extensions; python_version < '3.11'
 optimum
 peft
-pytest
 tensorflow-addons
 tf2onnx
 tf_slim
